Error handling

+++ dbio.py
@@ -2,8 +2,63 @@
 
 import sqlite3
 from bottle import response
+from datetime import datetime
 
+#*******************************************************************************
+def isInt(s):
+    try:
+        int(s)
+        return True
+    except:
+        return False
 
+#*******************************************************************************
+def validateDateFormat(datestr):
+    try:
+        print ("in try Y-m-d ")
+        print datestr
+        datetime.strptime(datestr, '%Y-%m-%d')
+        return True
+    except :
+        try:
+            print ("in try Y-m-d H:M ")
+            datetime.strptime(datestr, '%Y-%m-%d %H:%M')
+            return True
+        except:
+            try:
+                print ("in try Y-m-dTH:M ")
+                datetime.strptime(datestr, '%Y-%m-%dT%H:%M')
+                return True
+            except:
+                try:
+                    print ("in try Y-m-dTH:M:S")
+                    datetime.strptime(datestr, '%Y-%m-%dT%H:%M:%S')
+                    return True
+                except:
+                    try:
+                        print ("in try Y-m-d H:M:S")
+                        datetime.strptime(datestr, '%Y-%m-%d %H:%M:%S')
+                        return True
+                    except:
+                        return  False
+
+#*******************************************************************************
+def getAIDfromDate(date, pid):
+    stmt = "select aid \
+            from appointments \
+            where pid={0} and date=datetime('{1}')".format(pid, date)
+    r = dbCall(stmt)
+    print r
+    if not r:
+        return None
+    else:
+        return r[0]['aid']
+
+#*******************************************************************************
+def validateID( col, id, table):
+    stmt = "select count(*) as c from {0} where {1}={2}".format(table, col, id)
+    print stmt
+    return (dbCall(stmt)[0]['c'] == 1)
 
 #*******************************************************************************
 def dict_factory(cursor, row):
@@ -39,19 +94,25 @@ def stmtNextID(id, table):
     return r +1
 
 #*******************************************************************************
-def insAppointments(obj,pid):
-
+def insAppointments(apps,pid):
     try:
-        app = obj['appointments']
+        for date in apps:
+            print (date)
+            aid = stmtNextID('aid', 'appointments')
+            stmt = "insert into appointments(aid, pid, date) \
+                    select {0},{1}, datetime('{2}') \
+                    where not exists( \
+                    select 1,2,3 from appointments \
+                    where pid={1} and date=datetime('{2}'))".format(
+                    aid, pid, date)
+            print(stmt)
+            r = dbCall(stmt)
+            print(r)
     except:
         response.status = 400
         return  {'statuscode':400,
-                'reason': 'Syntax error or missing key \'appointments\'in your request'}
-    for date in app:
-        aid = stmtNextID('aid', 'appointments')
-        stmt = "insert into appointments values({0}, {1}, '{2}')".format(aid, pid, date)
-        dbCall(stmt)
-    pass
+                'reason': 'Syntax error datetime format'}
+
 
 #*******************************************************************************
 def postPoll(obj):
@@ -63,6 +124,7 @@ def postPoll(obj):
     try:
         n = obj['name']
         pw = obj['password']
+        apps = obj['appointments']
     except:
         response.status = 400
         return  {'statuscode':400,
@@ -71,9 +133,17 @@ def postPoll(obj):
         response.status = 400
         return {'statuscode':400,
                 'reason': 'the password is waaayyy tooo short. I mean: seriously?'}
+    #check, if dates are valid:
+    for d in apps:
+        if not validateDateFormat(d):
+            response.status = 400
+            return  {'statuscode':400,
+                    'reason': 'Date or datetime format is not valid'}
+
+    # if everything is valid, insertions can be done:
     stmt = "insert into polls values ({0}, '{1}', '{2}')".format(pid, pw, n)
     ret = dbCall(stmt)
-    insAppointments(obj, pid)
+    insAppointments(apps, pid)
 
     return showPoll(pid)
 
@@ -88,6 +158,16 @@ def putPoll(obj):
         response.status = 400
         return  {'statuscode':400,
                 'reason': 'Syntax error or missing key in your request'}
+    if  not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
+    #check, if dates are valid:
+    for d in apps:
+        if not validateDateFormat(d):
+            response.status = 400
+            return  {'statuscode':400,
+                    'reason': 'Date or datetime format is not valid'}
 
     # update polls table
     stmt = "update polls \
@@ -102,14 +182,24 @@ def putPoll(obj):
     stmt= "update appointments \
             set date='' \
             where pid={0} and date not in ({1})".format(pid, placeholders)
-    print(stmt)
-    print(apps)
-    dbCallMany(stmt, apps)
+    try:
+        dbCallMany(stmt, apps)
+    except:
+        response.status = 400
+        return  {'statuscode':400,
+                'reason': 'Syntax error: could not empty-string (verb) old appointments'}
+    # insert new appointments
+    insAppointments(apps, pid)
 
     return showPoll(pid)
 
 #*******************************************************************************
 def showPoll(pid):
+
+    if not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
     ### create return-data
     stmt = "select p.PID, p.name, a.aid, a.date  \
             from polls p, appointments a \
@@ -139,33 +229,63 @@ def showPoll(pid):
 #*******************************************************************************
 def postVote(obj):
     vid = stmtNextID('vid', 'votes')
-
+    obj['vid'] = vid
     pid = obj['pid']
-    date = obj['appointment']
-    stmt = "select aid \
-            from appointments \
-            where pid={0} and date='{1}'".format(pid, date)
-    print(stmt)
-    aid = dbCall(stmt)[0]['aid']
-    print(aid)
+    if not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
+    # get new date
+    try:
+        date = obj['appointment']
+    except:
+        response.status = 400
+        return  {'statuscode':400,
+                'reason': "Syntax error: could not find key 'appointment'"}
+
+    if not validateDateFormat(date):
+        response.status = 400
+        return {'statuscode':400,
+                'reason': "Date or datetime format is not valid"}
+
+    aid = getAIDfromDate(date, pid)
+    if aid is None:
+        response.status = 409
+        return {'statuscode':409,
+                'reason': "Date does not exist."}
 
     stmt = "insert into votes values({0}, {1}, {2})".format(vid, pid, aid)
     print(stmt)
-
     ret = dbCall(stmt)
     print(ret)
+
+    response.status = 201
+    return getVote(obj)
 
 #*******************************************************************************
 def deletePoll(obj):
 
-    pwd = obj['password']
     pid = obj['pid']
+    if not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
+    try:
+        pwd = obj['password']
+    except:
+        response.status = 400
+        return  {'statuscode':400,
+                'reason': 'Syntax error or missing password'}
 
     stmt = "select pid from polls where pid={0} and password='{1}'".format(pid, pwd)
     r = dbCall(stmt)
     print ("return pwd Call: {0}".format(r))
     # delete only, if query returned  at least one row
-    if r:
+    if r is None:
+        response.status = 403
+        return  {'statuscode':403,
+                'reason': 'password invalid, sry'}
+    else:
         stmt = "delete from votes where pid={0}".format(pid)
         ret = dbCall(stmt)
         print (ret)
@@ -177,11 +297,20 @@ def deletePoll(obj):
         stmt = "delete from polls where pid={0}".format(pid)
         ret = dbCall(stmt)
         print (ret)
+        response.status = 204 #empty body
 
 #*******************************************************************************
 def getVote(obj):
-    pid = obj['pid']
     vid = obj['vid']
+    if not isInt(vid) or not validateID('vid', vid, 'votes'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this VID does not exist.'"}
+    pid = obj['pid']
+    if not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
 
     stmt = "select a.date, v.vid \
             from appointments a, votes v\
@@ -193,26 +322,48 @@ def getVote(obj):
 #*******************************************************************************
 def deleteVote(obj):
         vid = obj['vid']
+        if not isInt(vid) or not validateID('vid', vid, 'votes'):
+            response.status = 404
+            return {'statuscode':404,
+                    'reason': "this VID does not exist.'"}
         pid = obj['pid']
+        if not isInt(pid) or not validateID('pid', pid, 'polls'):
+            response.status = 404
+            return {'statuscode':404,
+                    'reason': "this PID does not exist.'"}
+
         stmt = "delete from votes where pid={0} and vid={1}".format(pid, vid)
         ret = dbCall(stmt)
         print (ret)
+        response.status = 204
         return ret
 
 #*******************************************************************************
 def putVote(obj):
-    vid = obj['vid']
-    pid = obj['pid']
-    date = obj['appointment']
-    stmt = "select aid \
-            from appointments \
-            where pid={0} and date='{1}'".format(pid, date)
-    print(stmt)
-    r = dbCall(stmt)
-    print(r)
-    aid = r[0]['aid']
-    print(aid)
+    try:
+        date = obj['appointment']
+    except:
+        response.status = 400
+        return  {'statuscode':400,
+                'reason': "Syntax error: could not find key 'appointment'"}
 
+    pid = obj['pid']
+    if not isInt(pid) or not validateID('pid', pid, 'polls'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this PID does not exist.'"}
+
+    vid = obj['vid']
+    if not isInt(vid) or not validateID('vid', vid, 'votes'):
+        response.status = 404
+        return {'statuscode':404,
+                'reason': "this VID does not exist.'"}
+
+    aid = getAIDfromDate(date,pid)
+    if aid is None:
+        response.status = 409
+        return {'statuscode':409,
+                'reason': "Date does not exist."}
 
     # update votes table
     stmt = "update votes \
@@ -220,5 +371,4 @@ def putVote(obj):
             where vid={1}".format(aid, vid )
     r = dbCall(stmt)
     print(r)
-
     return getVote(obj)
